requirements.txt
# Core server
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
python-multipart>=0.0.9

# Settings / validation
pydantic>=2.6
pydantic-settings>=2.2

# Auth (Microsoft Entra ID / OIDC)
authlib>=1.3
itsdangerous>=2.2

# RAG / vector store
chromadb>=0.5

# Web search
# Either 'ddgs' (preferred) or 'duckduckgo-search' (fallback)
ddgs>=6.1
duckduckgo-search>=6.1

# Document parsing
python-docx>=1.1
openpyxl>=3.1
python-pptx>=0.6
PyMuPDF>=1.24
pdfplumber>=0.11

# Optional: Access/QVD support
pyodbc>=5.0
qvd>=0.3
pandas>=2.2
pyarrow>=15.0
duckdb>=1.0.0

# Misc
psutil>=5.9

# Dev/test
pytest>=8.0
